put heatMap into a component
